Cleaned up move_until_collision function Transferred a lot of the code into separate functions, greatly increasing readability.
Also added flags to check for exact collision with another entity.
I consider this to be a bit of a hack, and am looking to find a
way to check for collision with multiple entities at the same time
while preserving the returning of the colliding side for each one.

Maybe add a struct Collision that can hold A) an array of types of
entities to check, or B) an array of entities to check as well as
an additional array that holds collision information for every
affected entity.